+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # NEXT RELEASE
+* new `DocPrinter` class manages formatted document printing
 * `MessengerLogPdf` class for unstructured text message PDFs
 * `DocCfg` add `date_uncertain`, `is_valid_for_name_scan`, and `non_participants` fields
 * `Contact` now has `is_emailer`, `is_interesting`, `match_partial`, and `link_to_bio` props and `bio()` method for biographical panel